requirements.txt
av==11.0.0
contextily==1.5.0
exifread==3.0.0
ffmpeg==1.4
folium==0.15.0
fqdn==1.5.1
geodatasets==2023.12.0
geopandas==0.14.1
gpsphoto==2.2.3
gpxpy==1.6.1
image==1.5.33
isoduration==20.11.0
jsonpointer==2.4
jupyter==1.0.0
kaleido==0.2.1
mapclassify==2.6.1
moviepy==1.0.3
piexif==1.1.3
pip-chill==1.0.3
plotly==5.18.0
pyexiftool==0.5.6
pysocks==1.7.1
selenium==4.15.2
timm==0.9.10
transformers==4.35.0
uri-template==1.3.0
webcolors==1.13
yolov5==7.0.13
